get_xform_transform moved to Stage module
