+++ extern/usdhydra/utils.cpp
@@ -3,12 +3,7 @@
 
 #include <boost/random.hpp>
 
-#include <pxr/usd/usd/prim.h>
-#include <pxr/usd/usd/stage.h>
-#include <pxr/usd/usdGeom/xform.h>
-
 #include "utils.h"
-#include "stage.h"
 
 namespace usdhydra {
 
@@ -121,65 +116,9 @@ static PyObject *get_temp_dir_func(PyObject * /*self*/, PyObject * /*args*/)
   return PyUnicode_FromString(path.u8string().c_str());
 }
 
-static PyObject *get_xform_transform_func(PyObject * /*self*/, PyObject *args)
-{
-  long stageId;
-  const char *path;
-  if (!PyArg_ParseTuple(args, "ls", &stageId, &path)) {
-    Py_RETURN_NONE;
-  }
-
-  UsdStageRefPtr stage = stageCache->Find(UsdStageCache::Id::FromLongInt(stageId));
-  UsdPrim usd_prim = stage->GetPrimAtPath(SdfPath(path));
-  UsdGeomXform xForm = UsdGeomXform(usd_prim);
-
-  GfMatrix4d transform;
-  bool resets_xform_stack;
-
-  xForm.GetLocalTransformation(&transform, &resets_xform_stack);
-
-  auto matrix = (double(*)[4][4])transform.GetArray();
-
-  PyObject *pyMatrix = PyTuple_New(4);
-
-  PyObject *pyRow_0 = PyTuple_New(4);
-  PyObject *pyRow_1 = PyTuple_New(4);
-  PyObject *pyRow_2 = PyTuple_New(4);
-  PyObject *pyRow_3 = PyTuple_New(4);
-
-  PyTuple_SetItem(pyRow_0, 0, PyFloat_FromDouble((*matrix)[0][0]));
-  PyTuple_SetItem(pyRow_0, 1, PyFloat_FromDouble((*matrix)[0][1]));
-  PyTuple_SetItem(pyRow_0, 2, PyFloat_FromDouble((*matrix)[0][2]));
-  PyTuple_SetItem(pyRow_0, 3, PyFloat_FromDouble((*matrix)[0][3]));
-                                                         
-  PyTuple_SetItem(pyRow_1, 0, PyFloat_FromDouble((*matrix)[1][0]));
-  PyTuple_SetItem(pyRow_1, 1, PyFloat_FromDouble((*matrix)[1][1]));
-  PyTuple_SetItem(pyRow_1, 2, PyFloat_FromDouble((*matrix)[1][2]));
-  PyTuple_SetItem(pyRow_1, 3, PyFloat_FromDouble((*matrix)[1][3]));
-                                                         
-  PyTuple_SetItem(pyRow_2, 0, PyFloat_FromDouble((*matrix)[2][0]));
-  PyTuple_SetItem(pyRow_2, 1, PyFloat_FromDouble((*matrix)[2][1]));
-  PyTuple_SetItem(pyRow_2, 2, PyFloat_FromDouble((*matrix)[2][2]));
-  PyTuple_SetItem(pyRow_2, 3, PyFloat_FromDouble((*matrix)[2][3]));
-                                                         
-  PyTuple_SetItem(pyRow_3, 0, PyFloat_FromDouble((*matrix)[3][0]));
-  PyTuple_SetItem(pyRow_3, 1, PyFloat_FromDouble((*matrix)[3][1]));
-  PyTuple_SetItem(pyRow_3, 2, PyFloat_FromDouble((*matrix)[3][2]));
-  PyTuple_SetItem(pyRow_3, 3, PyFloat_FromDouble((*matrix)[3][3]));
-
-  PyTuple_SetItem(pyMatrix, 0, pyRow_0);
-  PyTuple_SetItem(pyMatrix, 1, pyRow_1);
-  PyTuple_SetItem(pyMatrix, 2, pyRow_2);
-  PyTuple_SetItem(pyMatrix, 3, pyRow_3);
-
-  return pyMatrix;
-}
-
-
 static PyMethodDef methods[] = {
   {"get_temp_file", get_temp_file_func, METH_VARARGS, ""},
   {"get_temp_dir", get_temp_dir_func, METH_VARARGS, ""},
-  {"get_xform_transform", get_xform_transform_func, METH_VARARGS, ""},
   {NULL, NULL, 0, NULL},
 };
